Sentry: Fix errors not always being caught
Occasionally MediaWiki will throw an Error. Sentry expected Exceptions, not Errors, so Sentry's code then flapped about and threw an Exception.

So Errors never made their way to /dev/stderr, or to Sentry.

The latest commit in the Sentry plugin removes the Exception type checking, which is what we want!