Disable flow for new talk pages
Need to think about how to convert existing pages. There's a maintenance script to convert them to wikitext, but by "convert" it means spit out the wikitext to stdout – not actually make changes on the wiki. Script also doesn't work on its own, I had to patch bits of it to get it to work.

Likely possible to extend that script and also have it do the deletion of the flow page, and creation of a replacement page from that wikitext. Gonna have to think hard about that though, about how to identify Flow pages, and then think hard about whether its a good idea to run on live – being a destructive operation and all.